move start_ipython to _ipython_utils
so it can be reused in the scheduler
