+++ distributed/worker.py
@@ -8,7 +8,7 @@ from multiprocessing.pool import ThreadPool
 import os
 import pkg_resources
 import tempfile
-from threading import current_thread, Thread, Event
+from threading import current_thread
 from time import time
 from timeit import default_timer
 import shutil
@@ -618,67 +618,18 @@ class Worker(Server):
     def get_data(self, stream, keys=None):
         return {k: dumps(self.data[k]) for k in keys if k in self.data}
 
-    def _start_ipython(self):
-        from IPython import get_ipython
-        if get_ipython() is not None:
-            raise RuntimeError("Cannot start IPython, it's already running.")
-
-        from zmq.eventloop.ioloop import ZMQIOLoop
-        from ipykernel.kernelapp import IPKernelApp
-        # save the global IOLoop instance
-        # since IPython relies on it, but we are going to put it in a thread.
-        save_inst = IOLoop.instance()
-        IOLoop.clear_instance()
-        zmq_loop = ZMQIOLoop()
-        zmq_loop.install()
-
-        # start IPython, disabling its signal handlers that won't work due to running in a thread:
-        app = self._ipython_kernel = IPKernelApp.instance(log=logger)
-        # Don't connect to the history database
-        app.config.HistoryManager.hist_file = ':memory:'
-        # listen on all interfaces, so remote clients can connect:
-        app.ip = self.ip
-        # disable some signal handling, logging
-        noop = lambda : None
-        app.init_signal = noop
-        app.log_connection_info = noop
-
-        # start IPython in a thread
-        # initialization happens in the thread to avoid threading problems
-        # with the sqlite history
-        evt = Event()
-        def _start():
-            app.initialize([])
-            app.kernel.pre_handler_hook = noop
-            app.kernel.post_handler_hook = noop
-            app.kernel.start()
-            app.kernel.loop = IOLoop.instance()
-            # save self in the IPython namespace as 'worker'
-            app.kernel.shell.user_ns['worker'] = self
-            from tornado.ioloop import PeriodicCallback
-            pc = PeriodicCallback(lambda : print(os.getpid(), file=sys.__stdout__), 1000, io_loop=zmq_loop)
-            pc.start()
-            evt.set()
-            zmq_loop.start()
-
-        zmq_loop_thread = Thread(target=_start)
-        zmq_loop_thread.daemon = True
-        zmq_loop_thread.start()
-        self._zmq_thread = zmq_loop_thread
-        assert evt.wait(timeout=5), "IPython didn't start in a reasonable amount of time."
-
-        # put the global IOLoop instance back:
-        IOLoop.clear_instance()
-        save_inst.install()
-        return app
-
     def start_ipython(self, stream):
         """Start an IPython kernel
 
         Returns Jupyter connection info dictionary.
         """
+        from ._ipython_utils import start_ipython
         if self._ipython_kernel is None:
-            self._ipython_kernel = self._start_ipython()
+            self._ipython_kernel = start_ipython(
+                ip=self.ip,
+                ns={'worker': self},
+                log=logger,
+            )
         return self._ipython_kernel.get_connection_info()
 
     def upload_file(self, stream, filename=None, data=None, load=True):
